Safari-609.1.11.1/LayoutTests/media/video-main-content-allow-then-deny-expected.txt

RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForVideoRateChange,OverrideUserGestureRequirementForMainContent"))
EVENT(canplaythrough)
EVENT(playing)
RUN(video.style.visibility = "hidden")
EVENT(pause)
END OF TEST
